Move KinesisDataFetcher to a new file
